Implemented the review comments
