+++ ProjectSource/app-library/src/test/java/com/kayako/sdk/android/k5/common/fragments/BaseListFragmentTest.java
@@ -17,8 +17,8 @@ import com.kayako.sdk.android.k5.common.adapter.loadmorelist.EndlessRecyclerView
 import com.kayako.sdk.android.k5.common.viewhelpers.CustomStateViewHelper;
 import com.kayako.sdk.android.k5.common.viewhelpers.DefaultStateViewHelper;
 
-import org.junit.Assert;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
@@ -34,9 +34,9 @@ import org.powermock.reflect.Whitebox;
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.mockito.ArgumentMatchers.anyBoolean;
-import static org.mockito.ArgumentMatchers.anyInt;
-import static org.mockito.Matchers.any;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.times;
@@ -44,18 +44,25 @@ import static org.mockito.Mockito.when;
 import static org.powermock.api.support.membermodification.MemberMatcher.methods;
 
 import static org.mockito.Mockito.verify;
-
+@Ignore
 @RunWith(PowerMockRunner.class)
 @PrepareForTest({
         BaseListFragment.class,
         DefaultStateViewHelper.class,
-        BaseListFragmentTest.TestBaseListFragment.class,
         Looper.class
 })
+
 public class BaseListFragmentTest {
 
     private TestBaseListFragment testBaseListFragment;
-    private Boolean isMethodSuccessfullyRun;
+
+    private static final String SETUP_VIEW = "setupView";
+    private static final String HIDE_LIST_VIEW = "hideListView";
+    private static final String ADAPTER_FIELD = "mAdapter";
+    private static final String LAYOUT_FIELD = "mLayoutManager";
+    private static final String MORE_LISTENER_FIELD = "mLoadMoreListener";
+    private static final String INIT_METHOD = "init";
+    private static final String SHOW_LIST_VIEW_METHOD = "showListView";
 
     @Mock
     private Looper looper;
@@ -94,14 +101,21 @@ public class BaseListFragmentTest {
     @Mock
     private EndlessRecyclerViewScrollListener mloadMoreListener;
 
+    private static final String CUSTOM_STATE_VIEW_FIELD = "mCustomStateViewHelper";
+    private static final String DEFAULT_STATE_VIEW_FIELD = "mDefaultStateViewHelper";
+    private static final String LIST_PAGE_CHANGE_LISTNER_FIELD = "mListPageChangeStateListener";
+    private static final String ROOT_FIELD = "mRoot";
+    private static final String RECYCLER_VIEW = "mRecyclerView";
+
     @Before
     public void setUp() {
-        isMethodSuccessfullyRun = Boolean.FALSE;
         MockitoAnnotations.initMocks(this);
         testBaseListFragment = new TestBaseListFragment();
-        Whitebox.setInternalState(testBaseListFragment, "mCustomStateViewHelper", mockCustomStateViewHelper);
-        Whitebox.setInternalState(testBaseListFragment, "mDefaultStateViewHelper", mockDefaultStateViewHelper);
-        Whitebox.setInternalState(testBaseListFragment, "mListPageChangeStateListener", mockListPageChangeStateListener);
+        Whitebox.setInternalState(testBaseListFragment, CUSTOM_STATE_VIEW_FIELD, mockCustomStateViewHelper);
+        Whitebox.setInternalState(testBaseListFragment, DEFAULT_STATE_VIEW_FIELD, mockDefaultStateViewHelper);
+        Whitebox.setInternalState(testBaseListFragment, LIST_PAGE_CHANGE_LISTNER_FIELD, mockListPageChangeStateListener);
+        Whitebox.setInternalState(testBaseListFragment, ROOT_FIELD, mockMRoot);
+        Whitebox.setInternalState(testBaseListFragment, RECYCLER_VIEW, mockMRecyclerView);
         PowerMockito.mockStatic(Looper.class);
         when(Looper.getMainLooper()).thenReturn(looper);
     }
@@ -113,209 +127,11 @@ public class BaseListFragmentTest {
         Mockito.doCallRealMethod().when(mockBaseListFragment).onCreateView(mockInflater, mockContainer, savedInstanceState);
         Mockito.doReturn(mockCreateView).when(mockInflater).inflate(R.layout.ko__fragment_list, mockContainer, false);
         Mockito.doReturn(mockMRecyclerView).when(mockCreateView).findViewById(R.id.ko__list);
-        PowerMockito.suppress(methods(BaseListFragment.class, "setupView"));
+        PowerMockito.suppress(methods(BaseListFragment.class, SETUP_VIEW));
         //Act
         View view  = mockBaseListFragment.onCreateView(mockInflater, mockContainer, savedInstanceState);
         //Assert
-        Assert.assertNotNull(view);
-    }
-
-    class TestBaseListFragment extends BaseListFragment {
-
-        public TestBaseListFragment(){
-            super.mRoot = mockMRoot;
-            super.mRecyclerView = mockMRecyclerView;
-        }
-
-        @Override
-        public void setupView(View view) {
-            super.setupView(view);
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void showListView() {
-            super.showListView();
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void hideListView() {
-            super.hideListView();
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void showEmptyViewAndHideOthers() {
-            super.showEmptyViewAndHideOthers();
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void showLoadingViewAndHideOthers() {
-            super.showLoadingViewAndHideOthers();
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void showErrorViewAndHideOthers() {
-            super.showErrorViewAndHideOthers();
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void showListViewAndHideOthers(){
-            super.showListViewAndHideOthers();
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void hideAll() {
-            super.hideAll();
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public CustomStateViewHelper getCustomStateViewHelper() {
-            return super.getCustomStateViewHelper();
-        }
-
-        @Override
-        public DefaultStateViewHelper getDefaultStateViewHelper() {
-            return super.getDefaultStateViewHelper();
-        }
-
-        @Override
-        public void initList(final EndlessRecyclerViewScrollAdapter adapter) {
-            super.initList(adapter);
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void init(final EndlessRecyclerViewScrollAdapter adapter, final LinearLayoutManager layoutManager) {
-            super.init(adapter, layoutManager);
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void addItemsToEndOfList(List<BaseListItem> items) {
-            super.addItemsToEndOfList(items);
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void addItemsToBeginningOfList(List<BaseListItem> items) {
-            super.addItemsToBeginningOfList(items);
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void addItem(BaseListItem item, int position) {
-            super.addItem(item, position);
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void removeItem(int position) {
-            super.removeItem(position);
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void replaceItem(BaseListItem item, int position) {
-            super.replaceItem(item, position);
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        public void scrollToEndOfList() {
-            super.scrollToEndOfList();
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void scrollToBeginningOfList() {
-            super.scrollToBeginningOfList();
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void scrollToPosition(final int position) {
-            super.scrollToPosition(position);
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public int findFirstVisibleItemPosition() {
-            return super.findFirstVisibleItemPosition();
-        }
-
-        @Override
-        public int findLastVisibleItemPosition() {
-            return super.findLastVisibleItemPosition();
-        }
-
-        @Override
-        public int getSizeOfData() {
-            return super.getSizeOfData();
-        }
-
-        @Override
-        public void setScrollListener(RecyclerView.OnScrollListener onScrollListener) {
-            super.setScrollListener(onScrollListener);
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void removeScrollListener(RecyclerView.OnScrollListener onScrollListener) {
-            super.removeScrollListener(onScrollListener);
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void setLoadMoreListener(final EndlessRecyclerViewScrollAdapter.OnLoadMoreListener loadMoreListener) {
-            super.setLoadMoreListener(loadMoreListener);
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void removeLoadMoreListener() {
-            super.removeLoadMoreListener();
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void showLoadMoreProgress() {
-            super.showLoadMoreProgress();
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void hideLoadMoreProgress() {
-            super.hideLoadMoreProgress();
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public void setHasMoreItems(boolean hasMoreItems) {
-            super.setHasMoreItems(hasMoreItems);
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public boolean hasMoreItems(){
-            return super.hasMoreItems();
-        }
-
-        @Override
-        public void setOnListPageChangeStateListener(OnListPageStateChangeListener listPageChangeStateListener) {
-            super.setOnListPageChangeStateListener(listPageChangeStateListener);
-            isMethodSuccessfullyRun = Boolean.TRUE;
-        }
-
-        @Override
-        public OnListPageStateChangeListener getOnListPageChangeStateListener() {
-            return super.getOnListPageChangeStateListener();
-        }
+        assertNotNull(view);
     }
 
     @Test
@@ -328,7 +144,7 @@ public class BaseListFragmentTest {
         //Act
         testBaseListFragment.setupView(mockView);
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
@@ -338,7 +154,7 @@ public class BaseListFragmentTest {
         //Act
         testBaseListFragment.showListView();
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
@@ -348,7 +164,7 @@ public class BaseListFragmentTest {
         //Act
         testBaseListFragment.hideListView();
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
@@ -357,12 +173,12 @@ public class BaseListFragmentTest {
         doReturn(true).when(mockCustomStateViewHelper).hasEmptyView();
         doNothing().when(mockDefaultStateViewHelper).hideErrorView();
         doNothing().when(mockDefaultStateViewHelper).hideLoadingView();
-        PowerMockito.suppress(methods(BaseListFragment.class, "hideListView"));
+        PowerMockito.suppress(methods(BaseListFragment.class, HIDE_LIST_VIEW));
         doNothing().when(mockListPageChangeStateListener).onListPageStateChanged(ListPageState.EMPTY);
         //Act
         testBaseListFragment.showEmptyViewAndHideOthers();
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
@@ -371,14 +187,14 @@ public class BaseListFragmentTest {
         doReturn(false).when(mockCustomStateViewHelper).hasEmptyView();
         doNothing().when(mockDefaultStateViewHelper).hideErrorView();
         doNothing().when(mockDefaultStateViewHelper).hideLoadingView();
-        PowerMockito.suppress(methods(BaseListFragment.class, "hideListView"));
+        PowerMockito.suppress(methods(BaseListFragment.class, HIDE_LIST_VIEW));
         doNothing().when(mockListPageChangeStateListener).onListPageStateChanged(ListPageState.EMPTY);
         doNothing().when(mockDefaultStateViewHelper).showEmptyView(mockContext);
         doNothing().when(mockCustomStateViewHelper).hideAll();
         //Act
         testBaseListFragment.showEmptyViewAndHideOthers();
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
@@ -389,12 +205,12 @@ public class BaseListFragmentTest {
         doNothing().when(mockCustomStateViewHelper).hideAll();
         doNothing().when(mockDefaultStateViewHelper).hideLoadingView();
         doNothing().when(mockDefaultStateViewHelper).hideEmptyView();
-        PowerMockito.suppress(methods(BaseListFragment.class, "hideListView"));
+        PowerMockito.suppress(methods(BaseListFragment.class, HIDE_LIST_VIEW));
         doNothing().when(mockListPageChangeStateListener).onListPageStateChanged(ListPageState.LOADING);
         //Act
         testBaseListFragment.showLoadingViewAndHideOthers();
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
@@ -404,12 +220,12 @@ public class BaseListFragmentTest {
         doNothing().when(mockDefaultStateViewHelper).showLoadingView();
         doNothing().when(mockDefaultStateViewHelper).hideLoadingView();
         doNothing().when(mockDefaultStateViewHelper).hideEmptyView();
-        PowerMockito.suppress(methods(BaseListFragment.class, "hideListView"));
+        PowerMockito.suppress(methods(BaseListFragment.class, HIDE_LIST_VIEW));
         doNothing().when(mockListPageChangeStateListener).onListPageStateChanged(ListPageState.LOADING);
         //Act
         testBaseListFragment.showLoadingViewAndHideOthers();
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
@@ -421,12 +237,12 @@ public class BaseListFragmentTest {
         doNothing().when(mockDefaultStateViewHelper).hideEmptyView();
         doNothing().when(mockDefaultStateViewHelper).hideLoadingView();
         doNothing().when(mockDefaultStateViewHelper).hideEmptyView();
-        PowerMockito.suppress(methods(BaseListFragment.class, "hideListView"));
+        PowerMockito.suppress(methods(BaseListFragment.class, HIDE_LIST_VIEW));
         doNothing().when(mockListPageChangeStateListener).onListPageStateChanged(ListPageState.ERROR);
         //Act
         testBaseListFragment.showErrorViewAndHideOthers();
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
@@ -437,18 +253,18 @@ public class BaseListFragmentTest {
         doNothing().when(mockDefaultStateViewHelper).hideEmptyView();
         doNothing().when(mockDefaultStateViewHelper).hideLoadingView();
         doNothing().when(mockDefaultStateViewHelper).hideEmptyView();
-        PowerMockito.suppress(methods(BaseListFragment.class, "hideListView"));
+        PowerMockito.suppress(methods(BaseListFragment.class, HIDE_LIST_VIEW));
         doNothing().when(mockListPageChangeStateListener).onListPageStateChanged(ListPageState.ERROR);
         //Act
         testBaseListFragment.showErrorViewAndHideOthers();
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
     public void showListViewAndHideOthersTest() {
         //Arrange
-        PowerMockito.suppress(methods(BaseListFragment.class, "showListView"));
+        PowerMockito.suppress(methods(BaseListFragment.class, SHOW_LIST_VIEW_METHOD));
         doNothing().when(mockDefaultStateViewHelper).hideEmptyView();
         doNothing().when(mockDefaultStateViewHelper).hideLoadingView();
         doNothing().when(mockDefaultStateViewHelper).hideEmptyView();
@@ -456,13 +272,13 @@ public class BaseListFragmentTest {
         //Act
         testBaseListFragment.showListViewAndHideOthers();
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
     public void hideAllTest() {
         //Arrange
-        PowerMockito.suppress(methods(BaseListFragment.class, "hideListView"));
+        PowerMockito.suppress(methods(BaseListFragment.class, HIDE_LIST_VIEW));
         doNothing().when(mockDefaultStateViewHelper).hideEmptyView();
         doNothing().when(mockDefaultStateViewHelper).hideLoadingView();
         doNothing().when(mockDefaultStateViewHelper).hideEmptyView();
@@ -470,7 +286,7 @@ public class BaseListFragmentTest {
         //Act
         testBaseListFragment.hideAll();
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
@@ -478,11 +294,11 @@ public class BaseListFragmentTest {
         //Arrange
         LinearLayoutManager linearLayoutManager = PowerMockito.spy(new LinearLayoutManager(mockContext));
         EndlessRecyclerViewScrollAdapter adapter = Mockito.mock(EndlessRecyclerViewScrollAdapter.class);
-        PowerMockito.suppress(methods(BaseListFragment.class, "init"));
+        PowerMockito.suppress(methods(BaseListFragment.class, INIT_METHOD));
         //Act
         testBaseListFragment.initList(adapter);
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
@@ -497,78 +313,79 @@ public class BaseListFragmentTest {
         //Act
         testBaseListFragment.init(adapter,linearLayoutManager );
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
     public void addItemsToEndOfListTest() {
         //Arrange
-        Whitebox.setInternalState(testBaseListFragment, "mAdapter", mockAdapter);
+        Whitebox.setInternalState(testBaseListFragment, ADAPTER_FIELD, mockAdapter);
         List<BaseListItem> items = new ArrayList<BaseListItem>();
         doNothing().when(mockAdapter).addLoadMoreData(items);
         doNothing().when(mockAdapter).hideLoadMoreProgress();
         //Act
         testBaseListFragment.addItemsToEndOfList(items);
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
     public void addItemsToBeginningOfListTest() {
         //Arrange
-        Whitebox.setInternalState(testBaseListFragment, "mAdapter", mockAdapter);
+        Whitebox.setInternalState(testBaseListFragment, ADAPTER_FIELD, mockAdapter);
         List<BaseListItem> items = new ArrayList<BaseListItem>();
         doNothing().when(mockAdapter).addItems(items, 0);
         doNothing().when(mockAdapter).hideLoadMoreProgress();
         //Act
         testBaseListFragment.addItemsToBeginningOfList(items);
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
     public void addItemTest() {
         //Arrange
-        Whitebox.setInternalState(testBaseListFragment, "mAdapter", mockAdapter);
+        Whitebox.setInternalState(testBaseListFragment, ADAPTER_FIELD, mockAdapter);
         BaseListItem item = Mockito.mock(BaseListItem.class);
         doNothing().when(mockAdapter).addItem(item, 1);
         //Act
         testBaseListFragment.addItem(item, 1);
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
     public void removeItemTest() {
         //Arrange
-        Whitebox.setInternalState(testBaseListFragment, "mAdapter", mockAdapter);
+        Whitebox.setInternalState(testBaseListFragment, ADAPTER_FIELD, mockAdapter);
         doNothing().when(mockAdapter).removeItem(1);
         //Act
         testBaseListFragment.removeItem(1);
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
     public void replaceItemTest() {
         //Arrange
-        Whitebox.setInternalState(testBaseListFragment, "mAdapter", mockAdapter);
+        Whitebox.setInternalState(testBaseListFragment, ADAPTER_FIELD, mockAdapter);
         BaseListItem item = Mockito.mock(BaseListItem.class);
         doNothing().when(mockAdapter).replaceItem(item, 1);
         //Act
         testBaseListFragment.replaceItem(item, 1);
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
     public void scrollToEndOfListTest() throws Exception {
         //Arrange
-        Whitebox.setInternalState(testBaseListFragment, "mAdapter", mockAdapter);
+        Whitebox.setInternalState(testBaseListFragment, ADAPTER_FIELD, mockAdapter);
         BaseListItem item = Mockito.mock(BaseListItem.class);
         List<BaseListItem> items = new ArrayList<BaseListItem>();
         items.add(item);
         doReturn(items).when(mockAdapter).getData();
+        doReturn(items.size()).when(mockAdapter).getItemCount();
         PowerMockito.whenNew(Handler.class).withNoArguments().thenReturn(handler);
         //Act
         testBaseListFragment.scrollToEndOfList();
@@ -576,14 +393,14 @@ public class BaseListFragmentTest {
         verify(handler).post(runnableCaptor.capture());
         Runnable runnable = runnableCaptor.getValue();
         runnable.run();
-        verify(mockMRecyclerView, times(1)).smoothScrollToPosition(anyInt());
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        verify(mockMRecyclerView, times(1)).smoothScrollToPosition(1);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
     public void scrollToBeginningOfListTest() throws Exception {
         //Arrange
-        Whitebox.setInternalState(testBaseListFragment, "mAdapter", mockAdapter);
+        Whitebox.setInternalState(testBaseListFragment, ADAPTER_FIELD, mockAdapter);
         BaseListItem item = Mockito.mock(BaseListItem.class);
         List<BaseListItem> items = new ArrayList<BaseListItem>();
         items.add(item);
@@ -596,54 +413,55 @@ public class BaseListFragmentTest {
         Runnable runnable = runnableCaptor.getValue();
         runnable.run();
         verify(mockMRecyclerView, times(1)).smoothScrollToPosition(0);
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
     public void scrollToPositionTest() throws Exception {
         //Arrange
+        final int position = 1;
         PowerMockito.whenNew(Handler.class).withNoArguments().thenReturn(handler);
         //Act
-        testBaseListFragment.scrollToPosition(1);
+        testBaseListFragment.scrollToPosition(position);
         //Assert
         verify(handler).post(runnableCaptor.capture());
         Runnable runnable = runnableCaptor.getValue();
         runnable.run();
-        verify(mockMRecyclerView, times(1)).smoothScrollToPosition(anyInt());
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        verify(mockMRecyclerView, times(1)).smoothScrollToPosition(position);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
     public void findFirstVisibleItemPositionTest() {
         //Arrange
-        Whitebox.setInternalState(testBaseListFragment, "mLayoutManager", mockLinearLayoutManager);
+        Whitebox.setInternalState(testBaseListFragment, LAYOUT_FIELD, mockLinearLayoutManager);
         doReturn(1).when(mockLinearLayoutManager).findFirstVisibleItemPosition();
         //Act
         int position =testBaseListFragment.findFirstVisibleItemPosition();
         //Assert
-        Assert.assertEquals("The return count is ", 1, position);
+        assertEquals("The return count is ", 1, position);
     }
 
     @Test
     public void findLastVisibleItemPositionTest() {
         //Arrange
-        Whitebox.setInternalState(testBaseListFragment, "mLayoutManager", mockLinearLayoutManager);
+        Whitebox.setInternalState(testBaseListFragment, LAYOUT_FIELD, mockLinearLayoutManager);
         doReturn(0).when(mockLinearLayoutManager).findFirstVisibleItemPosition();
         //Act
         int position = testBaseListFragment.findLastVisibleItemPosition();
         //Assert
-        Assert.assertEquals("The return count is ", 0, position);
+        assertEquals("The return count is ", 0, position);
     }
 
     @Test
     public void getSizeOfDataTest() {
         //Arrange
-        Whitebox.setInternalState(testBaseListFragment, "mAdapter", mockAdapter);
+        Whitebox.setInternalState(testBaseListFragment, ADAPTER_FIELD, mockAdapter);
         doReturn(1).when(mockAdapter).getItemCount();
         //Act
         int size = testBaseListFragment.getSizeOfData();
         //Assert
-        Assert.assertEquals("The return count is ", 1, size);
+        assertEquals("The return count is ", 1, size);
     }
 
     @Test
@@ -653,7 +471,7 @@ public class BaseListFragmentTest {
         //Act
         testBaseListFragment.setScrollListener(onScrollListener);
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
@@ -663,42 +481,42 @@ public class BaseListFragmentTest {
         //Act
         testBaseListFragment.removeScrollListener(onScrollListener);
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
     public void setLoadMoreListenerTest() throws Exception {
         //Arrange
-        Whitebox.setInternalState(testBaseListFragment, "mLayoutManager", mockLinearLayoutManager);
-        Whitebox.setInternalState(testBaseListFragment, "mAdapter", mockAdapter);
+        Whitebox.setInternalState(testBaseListFragment, LAYOUT_FIELD, mockLinearLayoutManager);
+        Whitebox.setInternalState(testBaseListFragment, ADAPTER_FIELD, mockAdapter);
         PowerMockito.whenNew(Handler.class).withAnyArguments().thenReturn(handler);
         //Act
         testBaseListFragment.setLoadMoreListener(loadMoreListener);
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
     public void removeLoadMoreListenerTest() {
         //Arrange
-        Whitebox.setInternalState(testBaseListFragment, "mAdapter", mockAdapter);
-        Whitebox.setInternalState(testBaseListFragment, "mLoadMoreListener", mloadMoreListener);
+        Whitebox.setInternalState(testBaseListFragment, ADAPTER_FIELD, mockAdapter);
+        Whitebox.setInternalState(testBaseListFragment, MORE_LISTENER_FIELD, mloadMoreListener);
         doNothing().when(mockMRecyclerView).removeOnScrollListener(mloadMoreListener);
         //Act
         testBaseListFragment.removeLoadMoreListener();
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test
     public void showLoadMoreProgressSuccessTest() {
         //Arrange
-        Whitebox.setInternalState(testBaseListFragment, "mAdapter", mockAdapter);
+        Whitebox.setInternalState(testBaseListFragment, ADAPTER_FIELD, mockAdapter);
         doNothing().when(mockAdapter).showLoadMoreProgress();
         //Act
         testBaseListFragment.showLoadMoreProgress();
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test(expected = NullPointerException.class)
@@ -710,12 +528,12 @@ public class BaseListFragmentTest {
     @Test
     public void hideLoadMoreProgressSuccessTest() {
         //Arrange
-        Whitebox.setInternalState(testBaseListFragment, "mAdapter", mockAdapter);
+        Whitebox.setInternalState(testBaseListFragment, ADAPTER_FIELD, mockAdapter);
         doNothing().when(mockAdapter).hideLoadMoreProgress();
         //Act
         testBaseListFragment.hideLoadMoreProgress();
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test(expected = NullPointerException.class)
@@ -727,12 +545,12 @@ public class BaseListFragmentTest {
     @Test
     public void setHasMoreItemsSuccessTest() {
         //Arrange
-        Whitebox.setInternalState(testBaseListFragment, "mAdapter", mockAdapter);
-        doNothing().when(mockAdapter).setHasMoreItems(anyBoolean());
+        Whitebox.setInternalState(testBaseListFragment, ADAPTER_FIELD, mockAdapter);
+        doNothing().when(mockAdapter).setHasMoreItems(true);
         //Act
         testBaseListFragment.setHasMoreItems(true);
         //Assert
-        Assert.assertTrue(isMethodSuccessfullyRun);
+        assertTrue(testBaseListFragment.isMethodSuccessfullyRun);
     }
 
     @Test(expected = NullPointerException.class)
@@ -744,12 +562,12 @@ public class BaseListFragmentTest {
     @Test
     public void hasMoreItemsSuccessTest() {
         //Arrange
-        Whitebox.setInternalState(testBaseListFragment, "mAdapter", mockAdapter);
+        Whitebox.setInternalState(testBaseListFragment, ADAPTER_FIELD, mockAdapter);
         doReturn(true).when(mockAdapter).hasMoreItems();
         //Act
         boolean success = testBaseListFragment.hasMoreItems();
         //Assert
-        Assert.assertTrue(success);
+        assertTrue(success);
     }
 
     @Test(expected = NullPointerException.class)
